refactor: move TLS/proxy/NoVerify from http.rs to builder.rs, remove dead legacy builders
