+++ src/common/mod.rs
@@ -5,6 +5,4 @@ pub use compression::{
     ContentEncoding, add_accept_encoding, add_content_encoding, compress_payload, decompress_body,
     decompress_stream, detect_encoding, detect_encodings,
 };
-pub use http::{
-    DEFAULT_POOL_MAX_IDLE_PER_HOST, HyperClient, build_hyper_client, build_hyper_client_with_pool,
-};
+pub use http::{DEFAULT_POOL_MAX_IDLE_PER_HOST, HyperClient};
