+++ src/webdav/builder.rs
@@ -4,11 +4,21 @@
 use anyhow::{Result, anyhow};
 use base64::Engine;
 use base64::engine::general_purpose::STANDARD as B64;
+use bytes::Bytes;
+use http_body_util::Full;
 use hyper::{Uri, header};
+use hyper_rustls::HttpsConnectorBuilder;
+use hyper_util::client::legacy::connect::proxy::Tunnel;
+use hyper_util::client::legacy::{Client, connect::HttpConnector};
+use hyper_util::rt::TokioExecutor;
+use rustls::pki_types::{CertificateDer, ServerName, UnixTime};
+use rustls::{ClientConfig, RootCertStore};
+use rustls_native_certs::load_native_certs;
+use std::sync::Arc;
 use std::time::Duration;
 use zeroize::Zeroize;
 
-use crate::common::http::{DEFAULT_POOL_MAX_IDLE_PER_HOST, build_hyper_client_full};
+use crate::common::http::{DEFAULT_POOL_MAX_IDLE_PER_HOST, HyperClient, MaybeProxied};
 use crate::webdav::client::{RequestCompressionMode, WebDavClient};
 
 /// Default request timeout applied when [`WebDavClientBuilder::timeout`] is
@@ -301,7 +311,7 @@ impl WebDavClientBuilder {
             None => None,
         };
 
-        let hyper_client = build_hyper_client_full(
+        let hyper_client = build_hyper_client(
             self.pool_max_idle_per_host,
             self.pool_idle_timeout,
             self.force_http1,
@@ -365,6 +375,177 @@ fn build_basic_auth_header(user: &str, pass: &str) -> Result<header::HeaderValue
     token.zeroize();
     val.zeroize();
     Ok(header_value?)
+}
+
+// ---------------------------------------------------------------------------
+// TLS configuration
+// ---------------------------------------------------------------------------
+
+/// A certificate verifier that accepts any server certificate.
+///
+/// # Warning
+///
+/// This completely disables TLS certificate verification. Only use in
+/// testing/debug scenarios — never in production.
+#[derive(Debug)]
+struct NoVerify;
+
+impl rustls::client::danger::ServerCertVerifier for NoVerify {
+    fn verify_server_cert(
+        &self,
+        _end_entity: &CertificateDer,
+        _intermediates: &[CertificateDer],
+        _server_name: &ServerName,
+        _ocsp_response: &[u8],
+        _now: UnixTime,
+    ) -> std::result::Result<rustls::client::danger::ServerCertVerified, rustls::Error> {
+        Ok(rustls::client::danger::ServerCertVerified::assertion())
+    }
+
+    fn verify_tls12_signature(
+        &self,
+        _message: &[u8],
+        _cert: &CertificateDer,
+        _dss: &rustls::DigitallySignedStruct,
+    ) -> std::result::Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
+        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
+    }
+
+    fn verify_tls13_signature(
+        &self,
+        _message: &[u8],
+        _cert: &CertificateDer,
+        _dss: &rustls::DigitallySignedStruct,
+    ) -> std::result::Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
+        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
+    }
+
+    fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
+        rustls::crypto::aws_lc_rs::default_provider()
+            .signature_verification_algorithms
+            .supported_schemes()
+    }
+}
+
+/// Build a rustls `ClientConfig` with native roots (fallback webpki),
+/// optional extra PEM trust roots, and optional danger mode.
+fn build_rustls_config(
+    extra_root_certs_pem: &[Vec<u8>],
+    danger_accept_invalid_certs: bool,
+) -> Result<ClientConfig> {
+    if danger_accept_invalid_certs {
+        #[cfg(debug_assertions)]
+        eprintln!(
+            "fast-dav-rs: WARNING — danger_accept_invalid_certs is enabled, \
+             TLS certificate verification is disabled"
+        );
+
+        let config = ClientConfig::builder()
+            .dangerous()
+            .with_custom_certificate_verifier(Arc::new(NoVerify))
+            .with_no_client_auth();
+        return Ok(config);
+    }
+
+    let mut roots = RootCertStore::empty();
+
+    match load_native_certs() {
+        result if !result.certs.is_empty() => {
+            for cert in result.certs {
+                let _ = roots.add(cert);
+            }
+        }
+        result => {
+            if !result.errors.is_empty() {
+                #[cfg(debug_assertions)]
+                eprintln!(
+                    "fast-dav-rs: falling back to webpki roots (native roots errors: {:?})",
+                    result.errors
+                );
+            }
+            roots.extend(webpki_roots::TLS_SERVER_ROOTS.iter().cloned());
+        }
+    }
+
+    for pem in extra_root_certs_pem {
+        for cert in rustls_pemfile::certs(&mut pem.as_slice()) {
+            let cert = cert.map_err(|e| anyhow!("failed to parse PEM certificate: {e}"))?;
+            let _ = roots.add(cert);
+        }
+    }
+
+    let config = ClientConfig::builder()
+        .with_root_certificates(roots)
+        .with_no_client_auth();
+    Ok(config)
+}
+
+// ---------------------------------------------------------------------------
+// Hyper client construction
+// ---------------------------------------------------------------------------
+
+/// Build a fully configured Hyper client.
+///
+/// Constructs the connector (with optional proxy tunnel), the TLS config
+/// (with optional extra roots / danger mode), and the Hyper client with
+/// pool settings. Called by [`WebDavClientBuilder::build`].
+#[allow(clippy::too_many_arguments)]
+fn build_hyper_client(
+    pool_max_idle_per_host: usize,
+    pool_idle_timeout: Option<Duration>,
+    force_http1: bool,
+    connect_timeout: Option<Duration>,
+    proxy: Option<Uri>,
+    proxy_basic_user: Option<String>,
+    proxy_basic_pass: Option<String>,
+    extra_root_certs_pem: &[Vec<u8>],
+    danger_accept_invalid_certs: bool,
+) -> Result<HyperClient> {
+    let mut http = HttpConnector::new();
+    http.enforce_http(false);
+    if let Some(t) = connect_timeout {
+        http.set_connect_timeout(Some(t));
+    }
+
+    let inner = match proxy {
+        Some(proxy_uri) => {
+            let mut tunnel = Tunnel::new(proxy_uri, http);
+            if let (Some(user), Some(pass)) = (proxy_basic_user, proxy_basic_pass) {
+                let basic = B64.encode(format!("{user}:{pass}"));
+                tunnel = tunnel.with_auth(
+                    format!("Basic {basic}")
+                        .parse()
+                        .map_err(|e| anyhow!("invalid proxy auth header: {e}"))?,
+                );
+            }
+            MaybeProxied::Tunneled(tunnel)
+        }
+        None => MaybeProxied::Direct(http),
+    };
+
+    let tls = build_rustls_config(extra_root_certs_pem, danger_accept_invalid_certs)?;
+
+    let https_builder = HttpsConnectorBuilder::new()
+        .with_tls_config(tls)
+        .https_or_http()
+        .enable_http1();
+
+    let https = if force_http1 {
+        https_builder.wrap_connector(inner)
+    } else {
+        https_builder.enable_http2().wrap_connector(inner)
+    };
+
+    let mut builder = Client::builder(TokioExecutor::new());
+    if !force_http1 {
+        builder.http2_adaptive_window(true);
+    }
+    builder.pool_max_idle_per_host(pool_max_idle_per_host);
+    if let Some(t) = pool_idle_timeout {
+        builder.pool_idle_timeout(t);
+    }
+
+    Ok(builder.build::<_, Full<Bytes>>(https))
 }
 
 #[cfg(test)]
